+++ DSCResources/Bitlocker/Bitlocker.schema.psm1
@@ -1,0 +1,145 @@
+<#
+.SYNOPSIS
+tbd
+
+.DESCRIPTION
+tbd
+
+.COMPONENT
+Information about PowerShell Modules to be required.
+Powershell Module: PSDesiredStateConfiguration
+Powershell Module: AuditPolicyDSC
+Powershell Module: SecurityPolicyDSC
+Powershell Module: GPRegistryPolicyDsc
+
+.EXAMPLE
+Get-DscResource -Module XOAPW11SecurityBaseline22H2DSC
+
+.NOTES
+For more information about advanced functions, call Get-Help with any
+of the topics in the links listed below.
+Module Name    : XOAPW11SecurityBaseline22H2DSC
+Resource Name  : Bitlocker.schema.psm1
+Author         : info@XOAP.io
+
+.LINK
+https://www.microsoft.com/en-us/download/details.aspx?id=55319
+
+.LINK
+https://www.powershellgallery.com/packages/BaselineManagement/2.9.0
+
+.LINK
+https://gallery.technet.microsoft.com/scriptcenter/PowerShellAccessControl-d3be7b83
+
+.LINK
+https://www.powershellgallery.com/packages/WindowsDefender/1.0.0.4
+#>
+
+Configuration Windows_11_v22H2_Security_Baseline_Bitlocker
+{
+
+     Import-DSCResource -ModuleName 'GPRegistryPolicyDsc' -ModuleVersion '1.2.0'
+     Import-DSCResource -ModuleName 'AuditPolicyDSC' -ModuleVersion '1.4.0.0'
+     Import-DSCResource -ModuleName 'SecurityPolicyDSC' -ModuleVersion '2.10.0.0'
+
+	Node Windows_11_v22H2_Security_Baseline_Bitlocker
+	{
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\FVE\UseEnhancedPin'
+         {
+              ValueName = 'UseEnhancedPin'
+              ValueData = 1
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\FVE'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\FVE\RDVDenyCrossOrg'
+         {
+              ValueName = 'RDVDenyCrossOrg'
+              ValueData = 0
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\FVE'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\FVE\DisableExternalDMAUnderLock'
+         {
+              ValueName = 'DisableExternalDMAUnderLock'
+              ValueData = 1
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\FVE'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\Power\PowerSettings\abfc2519-3608-4c2a-94ea-171b0ed546ab\DCSettingIndex'
+         {
+              ValueName = 'DCSettingIndex'
+              ValueData = 0
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Power\PowerSettings\abfc2519-3608-4c2a-94ea-171b0ed546ab'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\Power\PowerSettings\abfc2519-3608-4c2a-94ea-171b0ed546ab\ACSettingIndex'
+         {
+              ValueName = 'ACSettingIndex'
+              ValueData = 0
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Power\PowerSettings\abfc2519-3608-4c2a-94ea-171b0ed546ab'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClasses'
+         {
+              ValueName = 'DenyDeviceClasses'
+              ValueData = 1
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClassesRetroactive'
+         {
+              ValueName = 'DenyDeviceClassesRetroactive'
+              ValueData = 1
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions'
+         }
+
+         <#RegistryPolicyFile 'DELVALS_\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClasses'
+         {
+              ValueName = ''
+              Exclusive = $True
+              ValueData = ''
+              Ensure = 'Present'
+              ValueType = 'String'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClasses'
+         }#>
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClasses\1'
+         {
+              ValueName = '1'
+              ValueData = '{d48179be-ec20-11d1-b6b8-00c04fa372a7}'
+              ValueType = 'String'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\SOFTWARE\Policies\Microsoft\Windows\DeviceInstall\Restrictions\DenyDeviceClasses'
+         }
+
+         RegistryPolicyFile 'Registry(POL): HKLM:\System\CurrentControlSet\Policies\Microsoft\FVE\RDVDenyWriteAccess'
+         {
+              ValueName = 'RDVDenyWriteAccess'
+              ValueData = 1
+              ValueType = 'Dword'
+              TargetType = 'ComputerConfiguration'
+              Key = 'HKLM:\System\CurrentControlSet\Policies\Microsoft\FVE'
+         }
+
+         RefreshRegistryPolicy 'ActivateClientSideExtension'
+         {
+             IsSingleInstance = 'Yes'
+         }
+     }
+}
+