+++ Examples/dsc_configuration.ps1
@@ -1,0 +1,27 @@
+configuration default
+{
+    param
+    (
+        # Target nodes to apply the configuration
+        [string[]]$NodeName = 'localhost'
+    )
+
+    Import-Module XOAPW11SecurityBaseline22H2DSC
+    Import-DSCResource -ModuleName XOAPW11SecurityBaseline22H2DSC
+
+    Node $NodeName
+    {
+
+        $moduleRoot = [io.path]::GetDirectoryName((Get-Module XOAPW11SecurityBaseline22H2DSC).Path)
+        $examples = "$moduleRoot\Examples"
+
+        # Install the IIS role
+        WindowsFeature IIS
+        {
+            Ensure          = "Present"
+            Name            = "Web-Server"
+        }
+
+    }
+}
+