+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.1] - 2023-07-26
 
 ### Added
 
-- None
+- intial commit with Bitlocker, Computer, CredentialGuard, Defender Antivirus, Domain Security and IE11 resources
+- added module.test.ps1 for module compilation test
 
 ### Changed